% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{filter_data}
\alias{filter_data}
\title{Filter data based on percentage mitochondrial and total number of genes expressed}
\usage{
filter_data(sce)
}
\arguments{
\item{sce}{SingleCellExperiment}
}
\value{

}
\description{
Filter data based on percentage mitochondrial and total number of genes expressed
}
