% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10x_pbmc_10k.R
\name{get_10x_pbmc_10k}
\alias{get_10x_pbmc_10k}
\title{Get PBMC scRNA-seq data published in https://support.10xgenomics.com/single-cell-gene-expression/datasets/3.0.0/pbmc_10k_protein_v3  and annotated by https://www.biorxiv.org/content/10.1101/2021.05.20.445014v1}
\usage{
get_10x_pbmc_10k(cache_path)
}
\value{

}
\description{
Get PBMC scRNA-seq data published in https://support.10xgenomics.com/single-cell-gene-expression/datasets/3.0.0/pbmc_10k_protein_v3  and annotated by https://www.biorxiv.org/content/10.1101/2021.05.20.445014v1
}
