% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10x_sorted.R
\name{get_10x_sorted}
\alias{get_10x_sorted}
\title{Get sorted subsets of PBMC data published in https://www.nature.com/articles/ncomms14049}
\usage{
get_10x_sorted(cache_path)
}
\value{

}
\description{
Get sorted subsets of PBMC data published in https://www.nature.com/articles/ncomms14049
}
