% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10x_sorted_2017.R
\name{get_10x_sorted_2017}
\alias{get_10x_sorted_2017}
\title{Get 3' PBMC CITE-seq data published in https://www.biorxiv.org/content/10.1101/2020.10.12.335331v1}
\usage{
get_10x_sorted_2017(cache_path)
}
\value{

}
\description{
Get 3' PBMC CITE-seq data published in https://www.biorxiv.org/content/10.1101/2020.10.12.335331v1
}
