% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blish_2020.R
\name{get_blish_2020}
\alias{get_blish_2020}
\title{Get Healthy PBMC scRNA-seq data from https://www.nature.com/articles/s41591-020-0944-y and available https://www.covid19cellatlas.org/}
\usage{
get_blish_2020(cache_path)
}
\value{

}
\description{
Get Healthy PBMC scRNA-seq data from https://www.nature.com/articles/s41591-020-0944-y and available https://www.covid19cellatlas.org/
}
