% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ding_2019.R
\name{get_ding_2019}
\alias{get_ding_2019}
\title{Get PBMC scRNA-seq data from https://www.biorxiv.org/content/10.1101/632216v2.full.pdf and available https://singlecell.broadinstitute.org/single_cell/study/SCP424/single-cell-comparison-pbmc-data#/}
\usage{
get_ding_2019(cache_path)
}
\value{

}
\description{
Get PBMC scRNA-seq data from https://www.biorxiv.org/content/10.1101/632216v2.full.pdf and available https://singlecell.broadinstitute.org/single_cell/study/SCP424/single-cell-comparison-pbmc-data#/
}
