% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haniffa_2021.R
\name{get_haniffa_2021}
\alias{get_haniffa_2021}
\title{Get Healthy PBMC scRNA-seq data from https://www.nature.com/articles/s41591-020-0944-y and available https://www.covid19cellatlas.org/}
\usage{
get_haniffa_2021(cache_path)
}
\value{

}
\description{
Get Healthy PBMC scRNA-seq data from https://www.nature.com/articles/s41591-020-0944-y and available https://www.covid19cellatlas.org/
}
