% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hao_2020.R
\name{get_hao_2020}
\alias{get_hao_2020}
\title{Get 3' PBMC CITE-seq data published in https://www.biorxiv.org/content/10.1101/2020.10.12.335331v1}
\usage{
get_hao_2020(cache_path)
}
\value{

}
\description{
Get 3' PBMC CITE-seq data published in https://www.biorxiv.org/content/10.1101/2020.10.12.335331v1
}
