% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kotliarov_2020.R
\name{get_kotliarov_2020}
\alias{get_kotliarov_2020}
\title{Get PBMC CITE-seq data published in https://doi.org/10.1038/s41591-020-0769-8}
\usage{
get_kotliarov_2020(cache_path)
}
\value{

}
\description{
Get PBMC CITE-seq data published in https://doi.org/10.1038/s41591-020-0769-8
}
