% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/su_2020.R
\name{get_su_2020}
\alias{get_su_2020}
\title{Get PBMC CITE-seq data published in https://www.sciencedirect.com/science/article/pii/S0092867420314446 and annotated by https://www.biorxiv.org/content/10.1101/2021.05.20.445014v1}
\usage{
get_su_2020(cache_path)
}
\value{

}
\description{
Get PBMC CITE-seq data published in https://www.sciencedirect.com/science/article/pii/S0092867420314446 and annotated by https://www.biorxiv.org/content/10.1101/2021.05.20.445014v1
}
