% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsang_2021.R
\name{get_tsang_2021}
\alias{get_tsang_2021}
\title{Get Healthy PBMC scRNA-seq data from https://www.nature.com/articles/s41591-020-0944-y and available https://cellxgene.cziscience.com/collections/ed9185e3-5b82-40c7-9824-b2141590c7f0}
\usage{
get_tsang_2021(cache_path)
}
\value{

}
\description{
Get Healthy PBMC scRNA-seq data from https://www.nature.com/articles/s41591-020-0944-y and available https://cellxgene.cziscience.com/collections/ed9185e3-5b82-40c7-9824-b2141590c7f0
}
