% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalize_gene}
\alias{normalize_gene}
\title{Normalize gene expression to log counts per 10,000}
\usage{
normalize_gene(data)
}
\arguments{
\item{data}{matrix with protein expression data}
}
\value{

}
\description{
Normalize gene expression to log counts per 10,000
}
