% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalize_protein}
\alias{normalize_protein}
\title{Normalize protein expression by scaling each feature by geometric mean}
\usage{
normalize_protein(data)
}
\arguments{
\item{data}{matrix with protein expression data}
}
\value{

}
\description{
Normalize protein expression by scaling each feature by geometric mean
}
